backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.121.0
uvicorn[standard]==0.38.0
pydantic==2.12.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Cloud SDKs
boto3==1.34.0
azure-storage-blob==12.19.0
google-cloud-storage==2.14.0

# ML and Data Science
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
tensorflow==2.15.0

# Streaming
kafka-python==2.0.2

# Security & Encryption
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
pyjwt==2.8.0

# Caching & Distributed Systems
redis==5.0.1
hiredis==2.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
